// src/app/shared/types/api-types/bire-item-dto.interface.ts
import { BireDocumentDTO } from './bire-document-dto.interface';

export interface BireItemDTO {
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  name?: string;
  itemType?: string;
  itemLevel?: string;
  attachingPart?: boolean;
  lruSru?: string;
  functionCode?: string;
  ippeNodeType?: string;
  lifeRestrictionType?: string;
  configurableArticle?: boolean;
  itemRemarks?: string;
  serializedPart?: boolean;
  referChapter?: string;
  referSection?: string;
  referSubject?: string;
  referSheet?: string;
  referMarks?: string;
  hasCustomizedData?: boolean;
  itemFamilyCode?: string;
  inventoryVisible?: boolean;
  unmFamilyCode?: string;
  unmStructureType?: string;
  unmVariantCode?: string;
  unmountCode?: string;
  fixChapter?: string;
  fixSection?: string;
  fixSubject?: string;
  fixSheet?: string;
  fixMarks?: string;
  nhaChapter?: string;
  nhaSection?: string;
  nhaSubject?: string;
  nhaSheet?: string;
  nhaMarks?: string;
  sapClassCode?: string;
  itemNumber?: string;
  isItemHeadOfGroup?: boolean;
  isClassItem?: boolean;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  // tslint:disable-next-line: no-any
  additionalData?: any;
  bireDocumentDTO?: BireDocumentDTO;
}