// src/app/shared/types/api-types/bire-task-item-dto.interface.ts
export interface BireTaskItemDTO {
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  versionNumber?: string;
  taskCode?: string;
  taskVersion?: string;
  taskItemTypeCode?: string;
  item?: string;
}